coscene-backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1

# LangGraph & VLM
langgraph==0.1.1
langchain==0.2.15
langchain-community==0.2.15
langchain-anthropic==0.1.16
anthropic==0.31.2

# USD Core
# usd-core==25.11
# usd2gltf==0.3.5

# Utils
python-dotenv==1.0.0
redis==5.0.1
httpx==0.26.0
python-multipart==0.0.6

# Evaluation
scikit-image==0.24.0
Pillow==10.4.0
matplotlib==3.9.0
pandas==2.2.0
pyyaml==6.0.2

# Dev
# pytest==7.4.3
# pytest-asyncio==0.23.2
# black==23.12.0
# ruff==0.1.9
